requirements.txt
flask
python-dotenv
openai
google-generativeai